+++ api/src/main/java/mod/chiselsandbits/api/multistate/accessor/identifier/IArrayBackedAreaShapeIdentifier.java
@@ -13,9 +13,10 @@ public interface IArrayBackedAreaShapeIdentifier extends IAreaShapeIdentifier
 
     /**
      * Gives access to the backing long array.
+     *
      * @return The backing long array.
      */
-    byte[] getBackingData();
+    long[] getBackingData();
 
     /**
      * The palette that is in use for this identifier.
